src/lambda_funcs/2_rds_writer/lawn_rds_writer/requirements.txt
psycopg2-binary==2.8.6
pandas==1.1.2
boto3
sqlalchemy